Add tabs on user page for achievments, friends and settings
