+++ client/src/assets/scss/_variables.scss
@@ -2,6 +2,7 @@ $white: #fff;
 $black: #000;
 $grey-100: #f5f5f5;
 $grey-200: #dfdfdf;
+$grey-400: #a2a2a2;
 $grey-800: #636363;
 
 $text-color: #eff0eb;
